Updated Background Image and Navbar
Replace the current background image with a new one that was more relevant to our client. Also created a navbar.css file to handle responsiveness for smaller browsers to make our app look and function cleaner. Small issue on search assessment that needs to be cleaned up for better responsiveness.