Fixed sidebar logic for popouts
Changed onclick functions for sidebar and assessmentArea to actual javascript functions, and bound the functions to the on relevant onclicks.

Fixed sidebar asset images to be more uniform.

Added -active versions of each icon image.

Changed the undefined (null) icon to also be a popout, with its own low, med, and high priorities for QOL.

Renamed image files to be more uniform

Removed unused images